extract identity model
